fix: when run away don't remove map entity

+++ src/damage_system.rs
@@ -1,5 +1,5 @@
 use specs::prelude::*;
-use super::{CombatStats, SufferDamage, Player, Name, gamelog::GameLog, RunState, Battle, Monster};
+use super::{CombatStats, SufferDamage, Player, Name, gamelog::BattleLog, RunState, Battle, Monster};
 
 pub struct DamageSystem {}
 
@@ -18,15 +18,17 @@ impl<'a> System<'a> for DamageSystem {
     }
 }
 
+// TODO: 色々混ざってるので分割する
 pub fn delete_the_dead(ecs : &mut World) {
     let mut dead : Vec<Entity> = Vec::new();
+    let mut win = false;
     // Using a scope to make the borrow checker happy
     {
         let combat_stats = ecs.read_storage::<CombatStats>();
         let players = ecs.read_storage::<Player>();
         let names = ecs.read_storage::<Name>();
         let entities = ecs.entities();
-        let mut log = ecs.write_resource::<GameLog>();
+        let mut log = ecs.write_resource::<BattleLog>();
         for (entity, stats) in (&entities, &combat_stats).join() {
             if stats.hp < 1 {
                 let player = players.get(entity);
@@ -37,6 +39,7 @@ pub fn delete_the_dead(ecs : &mut World) {
                             log.entries.push(format!("{} is dead", &victim_name.name));
                         }
                         dead.push(entity);
+                        win = true;
                     }
                     Some(_) => {
                         let mut runstate = ecs.write_resource::<RunState>();
@@ -47,13 +50,12 @@ pub fn delete_the_dead(ecs : &mut World) {
         }
     }
 
-    // TODO: map上entityの削除もする。現在は戦闘用しか削除してない
+    // HPが0になった戦闘entityの削除
     for victim in dead {
         ecs.delete_entity(victim).expect("Unable to delete");
     }
 
-    // 戦闘に勝利したらmap entityを削除する
-    // TODO: 逃げたときもmap entityが消えている
+    // 戦闘に勝利したらmap entityを削除する。逃げたときはmap entityを削除しない
     let mut want_remove_battle = false;
     let mut dead_map_entity : Vec<Entity> = Vec::new();
     {
@@ -61,11 +63,12 @@ pub fn delete_the_dead(ecs : &mut World) {
         let combat_stats = ecs.read_storage::<CombatStats>();
         let monster = ecs.read_storage::<Monster>();
 
-        if (&entities, &combat_stats, &monster).join().count() == 0 {
+        // 攻撃の結果敵が残ってないときは*勝利*
+        // 攻撃してなくて敵が残ってないときは*逃走*
+        if (&entities, &combat_stats, &monster).join().count() == 0 && win{
             let battle_ecs = ecs.read_storage::<Battle>();
 
             for battle in (&battle_ecs).join() {
-                // 残ってないときは戦闘終了
                 let mut runstate = ecs.write_resource::<RunState>();
                 *runstate = RunState::AwaitingInput;
                 dead_map_entity.push(battle.monster);
@@ -83,8 +86,10 @@ pub fn delete_the_dead(ecs : &mut World) {
     }
 
     {
-        for map_entity in dead_map_entity {
-            ecs.delete_entity(map_entity).expect("Unable to delete");
+        if want_remove_battle {
+            for map_entity in dead_map_entity {
+                ecs.delete_entity(map_entity).expect("Unable to delete");
+            }
         }
     }
 }
